refactor: 스타일 정리
