+++ backend/src/main/java/com/deliveranything/domain/review/controller/ApiV1ReviewController.java
@@ -1,12 +1,10 @@
-package com.deliveranything.domain.reviews.controller;
+package com.deliveranything.domain.review.controller;
 
-import com.deliveranything.domain.reviews.dto.ReviewCreateRequest;
-import com.deliveranything.domain.reviews.dto.ReviewCreateResponse;
-import com.deliveranything.domain.reviews.entity.Review;
-import com.deliveranything.domain.reviews.service.ReviewService;
+import com.deliveranything.domain.review.dto.ReviewCreateRequest;
+import com.deliveranything.domain.review.dto.ReviewCreateResponse;
+import com.deliveranything.domain.review.service.ReviewService;
 import com.deliveranything.global.common.ApiResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,6 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @RequestMapping("api/v1/reviews")
 public class ApiV1ReviewController {
+
   private final ReviewService reviewService;
 
   //리뷰 생성
