+++ backend/src/main/java/com/deliveranything/domain/review/dto/ReviewCreateRequest.java
@@ -1,0 +1,13 @@
+package com.deliveranything.domain.review.dto;
+
+import com.deliveranything.domain.review.enums.ReviewTargetType;
+
+public record ReviewCreateRequest(
+    int rating,
+    String comment,
+    String[] photoUrls,
+    ReviewTargetType targetType,
+    Long targetId
+) {
+
+}
