+++ backend/src/main/java/com/deliveranything/domain/review/repository/ReviewPhotoRepository.java
@@ -1,7 +1,7 @@
-package com.deliveranything.domain.reviews.repository;
+package com.deliveranything.domain.review.repository;
 
-import com.deliveranything.domain.reviews.entity.Review;
-import com.deliveranything.domain.reviews.entity.ReviewPhoto;
+import com.deliveranything.domain.review.entity.Review;
+import com.deliveranything.domain.review.entity.ReviewPhoto;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
